Make JoseHeader package-private
